Significant changes, esp. new 'Block' class & wiggle
Various changes including a new 'Block' class that is responsible for
holding details of the current block: Its shape and the variants of
this shape if it's rotated. The ability to 'wiggle' the rotated block
left or right by 1 pixel if it doesn't fit is added; this should make
rotation feel slightly smoother. Also several other less significant
changes including some renaming.